pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "quintet"
version = "0.1.0"
description = "Quintet Multi-Agent Orchestrator + Ultra Mode 2.0 + Math Mode 3.0"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Quintet Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "pydantic>=2.0",
]

[project.optional-dependencies]
# Tier 1: Math Mode Core
math = [
    "sympy>=1.12",
    "numpy>=1.24",
    "scipy>=1.10",
]

# LLM Fabric (for async backends)
llm = [
    "aiohttp>=3.8",
    "pyyaml>=6.0",
]

# Tier 2: Advanced packs
optimization = ["cvxpy>=1.4"]
stats = ["statsmodels>=0.14", "scikit-learn>=1.3"]
ml = ["torch>=2.0"]
probabilistic = ["numpyro>=0.13", "jax>=0.4"]

# Builder/API
api = [
    "fastapi>=0.100",
    "uvicorn>=0.23",
]

# Development
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
]

# All dependencies (explicit list to avoid self-reference)
all = [
    # math
    "sympy>=1.12",
    "numpy>=1.24",
    "scipy>=1.10",
    # llm
    "aiohttp>=3.8",
    "pyyaml>=6.0",
    # api
    "fastapi>=0.100",
    "uvicorn>=0.23",
    # dev
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
quintet = "quintet.cli:main"
quintet-api = "quintet.builder.api:run_server"

[project.urls]
"Homepage" = "https://github.com/quintet/quintet"
"Documentation" = "https://github.com/quintet/quintet/blob/main/docs/QUINTET_ULTRA_MODE_REPLICATOR.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["quintet*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W"]
ignore = ["E501"]

